+++ src/utils/asArray.js
@@ -1,0 +1,8 @@
+const asArray = object => {
+    if (!object) {
+        return [];
+    }
+    return typeof object.toArray === 'function' ? object.toArray() : object;
+};
+
+export default asArray;
