+++ src/utils/navigateTo.js
@@ -1,0 +1,7 @@
+import history from './history';
+
+const navigateTo = path => {
+    history.push(path);
+};
+
+export default navigateTo;
